+++ localflavor/br/forms.py
@@ -13,6 +13,7 @@ from django.forms.fields import Field, RegexField, CharField, Select
 from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
 
+from localflavor import DeprecatedPhoneNumber
 from .br_states import STATE_CHOICES
 
 phone_digits_re = re.compile(r'^(\d{2})[-\.]?(\d{4,5})[-\.]?(\d{4})$')
@@ -39,7 +40,7 @@ class BRZipCodeField(RegexField):
                                              max_length, min_length, *args, **kwargs)
 
 
-class BRPhoneNumberField(Field):
+class BRPhoneNumberField(DeprecatedPhoneNumber, Field):
     """
     A form field that validates input as a Brazilian phone number, that must
     be in either of the following formats: XX-XXXX-XXXX or XX-XXXXX-XXXX.