+++ localflavor/it/forms.py
@@ -12,6 +12,7 @@ from django.forms.fields import Field, RegexField, Select, CharField
 from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
 
+from localflavor import DeprecatedPhoneNumber
 from .it_province import PROVINCE_CHOICES
 from .it_region import REGION_CHOICES, REGION_PROVINCE_CHOICES
 from .util import vat_number_validation, ssn_validation
@@ -118,7 +119,7 @@ class ITVatNumberField(Field):
             raise ValidationError(self.error_messages['invalid'])
 
 
-class ITPhoneNumberField(CharField):
+class ITPhoneNumberField(DeprecatedPhoneNumber, CharField):
     """
     A form field that validates input as an Italian phone number. Will strip
     any +39 country prefix from the number.