+++ localflavor/pk/forms.py
@@ -11,6 +11,7 @@ from django.forms.fields import CharField, RegexField, Select
 from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
 
+from localflavor import DeprecatedPhoneNumber
 from .pk_states import STATE_CHOICES
 
 POSTCODE_DIGITS_RE = re.compile(r'^(\d{5})$')
@@ -30,7 +31,7 @@ class PKPostCodeField(RegexField):
         super(PKPostCodeField, self).__init__(POSTCODE_DIGITS_RE, *args, **kwargs)
 
 
-class PKPhoneNumberField(CharField):
+class PKPhoneNumberField(DeprecatedPhoneNumber, CharField):
     """
     A form field that validates input as an Pakistani phone number.