+++ localflavor/sg/forms.py
@@ -12,6 +12,7 @@ from django.forms.fields import CharField, RegexField
 from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
 
+from localflavor import DeprecatedPhoneNumber
 
 PHONE_DIGITS_RE = re.compile(r'^[689](\d{7})$')
 
@@ -35,7 +36,7 @@ class SGPostCodeField(RegexField):
         super(SGPostCodeField, self).__init__(r'^\d{6}$', *args, **kwargs)
 
 
-class SGPhoneNumberField(CharField):
+class SGPhoneNumberField(DeprecatedPhoneNumber, CharField):
     """
     A form field that validates input as a Singapore phone number.